app_/requirements.txt
asgiref==3.5.0
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
boto3==1.21.23
botocore==1.24.23
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
cycler==0.10.0
dal-admin-filters==1.1.0
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.12
django-autocomplete-light==3.9.1
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-cron==0.5.1
django-filter==2.4.0
django-import-export==2.6.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-redis==5.0.0
django-robots==5.0
django-sendgrid-v5==0.8.0
django-storages==1.12.3
django-tables2==2.4.0
django-tinymce==3.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.1
docopt==0.4.0
et-xmlfile==1.1.0
fonttools==4.28.5
future==0.18.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.7.1
mailchimp==2.0.10
MarkupPy==1.14
matplotlib==3.5.1
multitasking==0.0.9
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.3
pandas==1.4.2
Pillow==9.0.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-barcode==0.13.1
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.1
PyYAML==5.4.1
qrcode==7.3.1
redis==4.1.0
requests==2.27.1
s3transfer==0.5.2
sendgrid==6.9.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
yfinance==0.1.63

